# trackify/openapi.yaml
openapi: 3.1.0
info:
  title: Trackify
  description: Trackify GPS tracking server API documentation. To use the API you need to have a server instance. For testing purposes you can use one of free [demo servers](https://www.trackify.org/demo-server/). For production use you can install your own server or get a [subscription service](https://www.trackify.org/product/tracking-server/).
  contact:
    name: Trackify Support
    url: https://www.trackify.org/
    email: support@trackify.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '6.14.5'
servers:
  - url: https://demo.trackify.org/api
    description: Demo Server 1
  - url: https://demo2.trackify.org/api
    description: Demo Server 2
  - url: https://demo3.trackify.org/api
    description: Demo Server 3
  - url: https://demo4.trackify.org/api
    description: Demo Server 4
  - url: https://server.trackify.org/api
    description: Subscription Server
  - url: http://{host}:{port}/api
    description: Other Server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '8082'
          - '80'
        default: '8082'
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
...
  /password/reset:
    post:
      operationId: postPasswordReset
      summary: Send password reset email
      description: Always responds with `200` regardless of whether the email matches a user.
      tags:
        - Password
      security: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
        required: true
      responses:
        '200':
          description: OK
          content: {}
        '400':
          description: Bad request
          content: {}
  /password/update:
    post:
      operationId: postPasswordUpdate
      summary: Set a new password using a reset token
      tags:
        - Password
      security: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                password:
                  type: string
                  format: password
        required: true
      responses:
        '200':
          description: OK
          content: {}
        '404':
          description: User not found
          content: {}
...
  /server:
    get:
      operationId: getServer
      summary: Fetch Server information
      tags:
        - Server
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '400':
          description: Bad request
          content: {}
    put:
      operationId: putServer
      summary: Update Server information
      tags:
        - Server
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Server'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '401':
          description: Unauthorized
          content: {}
        '400':
          description: Bad request
          content: {}
  /server/geocode:
    get:
      operationId: getServerGeocode
      summary: Reverse geocode coordinates
      tags:
        - Server
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Reverse geocoding is not enabled
          content: {}
        '401':
          description: Unauthorized
          content: {}
  /server/timezones:
    get:
      operationId: getServerTimezones
      summary: Fetch available timezones
      tags:
        - Server
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '401':
          description: Unauthorized
          content: {}
  /server/file/{path}:
    post:
      operationId: postServerFilePath
      summary: Upload a server file
      tags:
        - Server
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
        required: true
      responses:
        '200':
          description: OK
          content: {}
        '400':
          description: Bad request
          content: {}
  /server/gc:
    get:
      operationId: getServerGc
      summary: Trigger garbage collection
      tags:
        - Server
      responses:
        '200':
          description: OK
          content: {}
        '401':
          description: Unauthorized
          content: {}
  /server/cache:
    get:
      operationId: getServerCache
      summary: Fetch cache diagnostics
      tags:
        - Server
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Unauthorized
          content: {}
  /server/reboot:
    post:
      operationId: postServerReboot
      summary: Reboot the server process
      tags:
        - Server
      responses:
        '200':
          description: OK
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '400':
          description: Bad request
          content: {}
...
  /events/{id}:
    get:
      operationId: getEventsId
      summary: Fetch an Event
      description: Returns a single Event record with full context for the specified identifier
      tags:
        - Events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          description: Unauthorized
          content: {}
        '404':
          description: Not found
          content: {}
...
  /reports/events/{type}:
    get:
      operationId: getReportsEventsType
      summary: Download the events report as a spreadsheet or send it by email
      description: Use `type=xlsx` to download the report, `type=mail` to deliver it by email asynchronously.
      tags:
        - Reports
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum:
              - xlsx
              - mail
        - name: deviceId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: groupId
          in: query
          style: form
          schema:
            type: array
            items:
              type: integer
        - name: type
          in: query
          description: Event types to include; `%` matches all
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
        - name: alarm
          in: query
          description: Alarm types to include
          style: form
          explode: false
          schema:
            type: array
            items:
              type: string
        - name: from
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Spreadsheet attachment
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '204':
          description: Report queued for email delivery
          content: {}
        '401':
          description: Unauthorized
          content: {}
        '404':
          description: Not found
          content: {}
...
components:
  schemas:
...
    Server:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique server configuration identifier
        registration:
          type: boolean
          description: Whether new user registrations are allowed
        readonly:
          type: boolean
          description: When true only administrators can modify server-wide settings
        deviceReadonly:
          type: boolean
          description: Disallow device attribute changes for non-admins
        limitCommands:
          type: boolean
          description: Restrict command execution to supported protocol commands
        map:
          type: string
          description: Default map layer identifier
        bingKey:
          type: string
          description: Bing Maps API key used when Bing is selected as a provider
        mapUrl:
          type: string
          description: Custom tile server URL template if configured
        poiLayer:
          type: string
          description: External point-of-interest layer configuration
        announcement:
          type: string
          description: Message displayed to all users in the web application
        latitude:
          type: number
          description: Default map center latitude
        longitude:
          type: number
          description: Default map center longitude
        zoom:
          type: integer
          description: Default map zoom level
        version:
          type: string
          description: Trackify server version string
        forceSettings:
          type: boolean
          description: Forces users to use the server-wide settings instead of their own
        coordinateFormat:
          type: string
          description: Default coordinate format for displaying positions
        openIdEnabled:
          type: boolean
          description: Indicates whether OpenID authentication is available
        openIdForce:
          type: boolean
          description: Require OpenID authentication for all users when enabled
        attributes:
          type: object
          description: Additional server-level configuration values
          properties: {}
...
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique event identifier
        type:
          type: string
          description: Event type name
        eventTime:
          type: string
          description: in ISO 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
        deviceId:
          type: integer
          format: int64
          description: Device associated with the event
        positionId:
          type: integer
          format: int64
          description: Position record linked to the event when applicable
        geofenceId:
          type: integer
          format: int64
          description: Geofence referenced by the event if any
        maintenanceId:
          type: integer
          description: Maintenance record tied to the event
        attributes:
          type: object
          properties: {}
          description: Additional event-specific attributes
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer